requirements.txt
python-telegram-bot==20.7
requests==2.25.1
python-dotenv==0.21.0
